fix bug and add latest comics api

+++ server/database/migrations/2022_10_02_100919_create_comic_episodes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
         Schema::create('comic_episodes', function (Blueprint $table) {
             $table->id();
             $table->foreignId('comic_id')->constrained();
-            $table->unsignedInteger('episode_number');
+            $table->float('episode_number');
 //            $table->date('published_date');
             $table->unsignedBigInteger('likes')->default(0);
             $table->unsignedBigInteger('views')->default(0);
